Extract BlocksInput component
